requirements.txt
kivy
pyjnius
plyer
pydantic
